+++ R/posterior_latents.R
@@ -43,7 +43,7 @@ posterior_infectious <-
   function(object, ...) UseMethod("posterior_infectious", object)
 
 #' @rdname posterior_infectious
-# @export
+#' @export
 posterior_infectious.epimodel <-
   function(object,
            newdata = NULL,