+++ inst/include/stan_meta_header.hpp
@@ -1,4 +1,0 @@
-#include "csr_matrix_times_vector2.hpp"
-
-
-